Split timer and component logic
